Use Bootstrap dropdown background color
Previously the dropdown did not have an explicit background color set, so it defaulted to the `body` background color from Select2, which is `#fff` . If you were using a dark theme such as [bootswatch/darkly](https://bootswatch.com/darkly/) , it would not match the theme backgrounds. 

This change brings the dropdown colors in line with Bootstrap's dropdown theme.